feat: tailcall-version workspace
