+++ Cargo.toml
@@ -50,6 +50,7 @@ opentelemetry-otlp = { version = "0.15.0", features = [
 ], optional = true }
 opentelemetry-system-metrics = { version = "0.1.8", optional = true }
 tailcall-http-cache = {path = "tailcall-http-cache", optional = true }
+tailcall-version = { path = "./tailcall-version", optional = true }
 
 
 # dependencies safe for wasm:
@@ -197,6 +198,7 @@ cli = [
     "dep:opentelemetry-system-metrics",
     "dep:tailcall-tracker",
     "dep:tailcall-http-cache",
+    "dep:tailcall-version"
 ]
 
 # Feature flag to enable all default features.
@@ -219,6 +221,7 @@ members = [
     "tailcall-tracker",
     "tailcall-hasher",
     "tailcall-http-cache",
+    "tailcall-version",
 ]
 
 # Boost execution_spec snapshot diffing performance
